+++ .env.example
@@ -4,6 +4,7 @@ BOT_TOKEN=
 
 SUGGESTIONS_CHANNEL=
 CORRECTIONS_CHANNEL=
+REPORTS_CHANNEL=
 COMMANDS_CHANNEL=
 REMINDERS_CHANNEL=
 LOGS_CHANNEL=
