+++ src/jokes.ts
@@ -1,6 +1,6 @@
 import { constants as fsConstants, promises as fs } from 'fs';
 import { Correction, Joke, Suggestion, Report } from './typings';
-import path from 'path';
+import { join } from 'path';
 
 import { AsyncQueue } from '@sapphire/async-queue';
 
@@ -20,7 +20,7 @@ class JokesLoader {
   }
 
   private async init() {
-    const jokesPath = path.join(__dirname, '../blagues.json');
+    const jokesPath = join(__dirname, '../blagues.json');
     try {
       await fs.access(jokesPath, fsConstants.R_OK | fsConstants.W_OK);
     } catch (error) {
@@ -37,7 +37,7 @@ class JokesLoader {
   public async mergeJoke(
     proposal: Correction | Suggestion
   ): Promise<{ success: boolean; joke_id?: number; error?: string }> {
-    const jokesPath = path.join(__dirname, '../blagues.json');
+    const jokesPath = join(__dirname, '../blagues.json');
     try {
       await fs.access(jokesPath, fsConstants.R_OK | fsConstants.W_OK);
     } catch (error) {
@@ -79,7 +79,7 @@ class JokesLoader {
   }
 
   public async removeJoke(proposal: Report): Promise<{ success: boolean; joke_id?: number; error?: string }> {
-    const jokesPath = path.join(__dirname, '../blagues.json');
+    const jokesPath = join(__dirname, '../blagues.json');
     try {
       await fs.access(jokesPath, fsConstants.R_OK | fsConstants.W_OK);
     } catch (error) {
